Rename: 디렉토리 분류 기준을 도메인 -> MVC 순서로 변경
