+++ src/main/java/ojosama/talkak/comment/controller/CommentController.java
@@ -1,6 +1,8 @@
-package ojosama.talkak.comment;
+package ojosama.talkak.comment.controller;
 
 import java.util.List;
+import ojosama.talkak.comment.dto.CommentResponse;
+import ojosama.talkak.comment.service.CommentService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
