+++ src/main/java/ojosama/talkak/comment/dto/CommentRequest.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.comment;
+package ojosama.talkak.comment.dto;
 
 public record CommentRequest(String content) {
