+++ src/main/java/ojosama/talkak/comment/dto/CommentResponse.java
@@ -1,6 +1,6 @@
-package ojosama.talkak.comment;
+package ojosama.talkak.comment.dto;
 
-import ojosama.talkak.user.MemberResponse;
+import ojosama.talkak.member.dto.MemberResponse;
 
 public record CommentResponse(long commentId, MemberResponse member, String content) {
