+++ src/main/java/ojosama/talkak/comment/model/Comment.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.model;
+package ojosama.talkak.comment.model;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -8,6 +8,8 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
+import ojosama.talkak.member.model.Member;
+import ojosama.talkak.video.model.Video;
 
 @Entity
 @Table(name = "comment")
