+++ src/main/java/ojosama/talkak/comment/repository/CommentRepository.java
@@ -1,7 +1,7 @@
-package ojosama.talkak.comment;
+package ojosama.talkak.comment.repository;
 
 import java.util.List;
-import ojosama.talkak.model.Comment;
+import ojosama.talkak.comment.model.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
