+++ src/main/java/ojosama/talkak/member/dto/MemberResponse.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.user;
+package ojosama.talkak.member.dto;
 
 public record MemberResponse(Long memberId, String imageUrl, String username) {
