+++ src/main/java/ojosama/talkak/member/model/Member.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.model;
+package ojosama.talkak.member.model;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
@@ -9,6 +9,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import java.util.List;
+import ojosama.talkak.comment.model.Comment;
 
 @Entity
 @Table(name = "member")
