+++ src/main/java/ojosama/talkak/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package ojosama.talkak.user;
+package ojosama.talkak.member.repository;
 
-import ojosama.talkak.model.Member;
+import ojosama.talkak.member.model.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
