+++ src/main/java/ojosama/talkak/reaction/repository/ReactionRepository.java
@@ -1,6 +1,7 @@
-package ojosama.talkak.reaction;
+package ojosama.talkak.reaction.repository;
 
-import ojosama.talkak.model.Reaction;
+import ojosama.talkak.reaction.model.Reaction;
+import ojosama.talkak.reaction.model.ReactionId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
