+++ src/main/java/ojosama/talkak/video/model/Video.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.model;
+package ojosama.talkak.video.model;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
@@ -11,6 +11,7 @@ import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import java.time.LocalDateTime;
 import java.util.List;
+import ojosama.talkak.comment.model.Comment;
 import org.hibernate.annotations.CreationTimestamp;
 
 @Entity
